tests/wpt/tests/css/cssom-view/historical.html

<!doctype html>
<meta charset="utf-8">
<title>Historical features</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  assert_false("ClientRectList" in self);
}, "Support for ClientRectList");

test(function() {
  assert_false("ClientRect" in self);
}, "Support for ClientRect");
</script>
